requirements.txt
pyodbc
python-dotenv
openai
streamlit
pandas
azure-storage-blob
pypdf
azure-ai-documentintelligence
azure-search-documents
sentence-transformers
streamlit_chat
msal
requests
numpy==1.26.4
pillow
azure-storage-queue
azure-cosmos
streamlit-aggrid
fastapi
fastapi-cors
uvicorn
pydantic
python-multipart
langchain-community
langchain_openai
langchain
